refactor: extract formatApiFailureMessage into shared lib/format-api-error.ts
Agent-Logs-Url: https://github.com/FriskyDevelopments/stixmagic-web/sessions/b6c46e89-0ff7-4b3d-a1c4-6af3bf968961
